From 9eb208c332e2ae3e0524e6cf2b6ecdfa25d2483c Mon Sep 17 00:00:00 2001 From: Irides <irides@irides.network> Date: Tue, 5 Apr 2022 15:39:34 -0500 Subject: [PATCH 14/20] build/plat: improve handling of get_override().
The existing functionality of get_override was poorly specified and
ill-purposed for boolean flags. This change extracts the core
variable retrieval logic to a helper function and adds a new handler
`get_override_flag` which special cases boolean flags.

The new behavior will also perform type checking on kwargs and inform
the user of the desired type expected.